+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - **Breaking Change**: `ParserBuilder.checkSymbols` is now private, use `options` in `ParserBuilder.build` instead.
   - **Breaking Change**: Remove `AdvancedBuilder.expand`, make `AdvancedBuilder` extend `ParserBuilder`, they both implement `IParserBuilder`.
   - Add `ParserBuilder.use` to apply custom decorators with chain call.
+  - Add `IParserBuilder.priority/leftSA/rightSA` for simplified conflict resolution.
 
 ## v0.7.0